+++ components/spider-derive/src/lib.rs
@@ -1,4 +1,4 @@
-mod enum_str;
+mod mysql;
 
 use proc_macro::TokenStream;
 use syn::{DeriveInput, parse_macro_input};
@@ -6,7 +6,7 @@ use syn::{DeriveInput, parse_macro_input};
 #[proc_macro_derive(QuotedEnumStr)]
 pub fn derive_quoted_enum_str(input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as DeriveInput);
-    enum_str::derive_quoted_enum_str(&input)
+    mysql::derive_quoted_enum_str(&input)
         .unwrap_or_else(syn::Error::into_compile_error)
         .into()
 }
